Fixed merges and enhanced tool to have two functionalities
