+++ .gitignore
@@ -5,5 +5,20 @@ __pycache__/
 # Ignore virtual env
 venv/
 Essential8_scanner/venv/
+
+# Ignore Virtual environments
+.venv/
+venv/
+env/
+
+# Ignore individual project outputs/reports
+scan_report.csv          
+reports_out/             
+previews/               
+
 # macOS
 .DS_Store
+.vscode/
+.idea/
+
+
