--- evidence/example.txt
@@ -1,1 +1,0 @@
-executable files
