+++ evidence/restrict_admin_privileges/Screenshot6 RPA RB-06.png
