+++ scan_report.csv
@@ -1,6 +1,6 @@
 UserID,Image,Strategy,TestID,Sub-Strategy,ML Level,Pass/Fail,Priority,Recommendation,Evidence Extract
-Camille,Capture5 Microsoft Endpoint Security Block EXE 01.jpeg,Application Control,ML1-AC-01,Executables Files must be blocked,1,Pass,High,(1) Ensure executable files (EXE and COM files) cannot be run from user or temp folders by standard users (2) Use AppLocker or Windows Defender Application Control rules,executable files; re:\bblock(?:ed|ing)?\b
-Camille,Capture4 AppLocker Properties.png,Application Control,ML1-AC-01 to ML1-AC-07,,1,Fail,High,(1) Use AppLocker or Windows Defender Application Control rules to block all listed file types from running in user or temporary folders. (2) Ensure rules are enforced on all workstations and reviewed periodically (ideally once per quarter).,
-Camille,Capture6 Windows Application Locker Configuration.png,Application Control,ML1-AC-01,Executables Files must be blocked,1,Pass,High,(1) Ensure executable files (EXE and COM files) cannot be run from user or temp folders by standard users (2) Use AppLocker or Windows Defender Application Control rules,executable rules; re:\b[\w-]+\s*\.\s*(?:exe|com)\b; re:\b[a-z]den[vy]\b
-Camille,Capture3 Unrelated.PNG,Application Control,ML1-AC-01 to ML1-AC-07,,1,Fail,High,(1) Use AppLocker or Windows Defender Application Control rules to block all listed file types from running in user or temporary folders. (2) Ensure rules are enforced on all workstations and reviewed periodically (ideally once per quarter).,
-
+Camille,Screenshot1 AC-01 03 04.png,Application Control,ML1-AC-01,Executables Files must be blocked,1,Pass,High,(1) Ensure executable files (EXE and COM files) cannot be run from user or temp folders by standard users (2) Use AppLocker or Windows Defender Application Control rules,executable rules; re:\benforc(?:e|ed|ing)\b
+Camille,Screenshot1 AC-01 03 04.png,Application Control,ML1-AC-03,Unauthorised Scripts must be blocked,1,Pass,High,"(1) Ensure script files (SH, BASH, CSH, KSH, BAT, CMD, JS, VBS, PY, PS1, PL, PHP, SCPT, WSF, OSTS) cannot be run from user or temp folders by standard users (2) Use AppLocker or Windows Defender Application Control rules",script rules; re:\benforc(?:e|ed|ing)\b
+Camille,Screenshot1 AC-01 03 04.png,Application Control,ML1-AC-04,Installers Files must be blocked,1,Pass,Medium,"(1) Ensure installer files (MSI, MSM, MSP, MST, IDT, CUB and PCP) cannot be run from user or temp folders by standard users (2) Use AppLocker or Windows Defender Application Control rules",windows installer rules; re:\benforc(?:e|ed|ing)\b
+Camille,Screenshot2 AC-02.webp,Application Control,ML1-AC-02,Software Library Files must be blocked,1,Pass,Medium,(1) Ensure software library files (DLL and OCX files) cannot be run from user or temp folders by standard users (2) Use AppLocker or Windows Defender Application Control rules,dll rules; re:\b(?:dll|ocx)\b; re:\benforc(?:e|ed|ing)\b
+Camille,Screenshot3 AC-01.jpeg,Application Control,ML1-AC-01,Executables Files must be blocked,1,Pass,High,(1) Ensure executable files (EXE and COM files) cannot be run from user or temp folders by standard users (2) Use AppLocker or Windows Defender Application Control rules,executable files; re:\bblock(?:ed|ing)?\b
